🚚 Move mixedCase arbitrary into /arbitrary
