+++ src/arbitrary/_internals/MixedCaseArbitrary.ts
@@ -1,24 +1,10 @@
 import { Random } from '../../random/generator/Random';
 import { Stream } from '../../stream/Stream';
-import { bigUintN } from '../../arbitrary/bigUintN';
-import { Arbitrary } from './definition/Arbitrary';
-import { NextArbitrary } from './definition/NextArbitrary';
-import { convertFromNext, convertToNext } from './definition/Converters';
-import { NextValue } from './definition/NextValue';
+import { bigUintN } from '../bigUintN';
+import { NextArbitrary } from '../../check/arbitrary/definition/NextArbitrary';
+import { convertToNext } from '../../check/arbitrary/definition/Converters';
+import { NextValue } from '../../check/arbitrary/definition/NextValue';
 import { makeLazy } from '../../stream/LazyIterableIterator';
-
-/**
- * Constraints to be applied on {@link mixedCase}
- * @remarks Since 1.17.0
- * @public
- */
-export interface MixedCaseConstraints {
-  /**
-   * Transform a character to its upper and/or lower case version
-   * @remarks Since 1.17.0
-   */
-  toggleCase?: (rawChar: string) => string;
-}
 
 /** @internal */
 export function countToggledBits(n: bigint): number {
@@ -57,7 +43,7 @@ type MixedCaseArbitraryContext = {
 };
 
 /** @internal */
-class MixedCaseArbitrary extends NextArbitrary<string> {
+export class MixedCaseArbitrary extends NextArbitrary<string> {
   constructor(
     private readonly stringArb: NextArbitrary<string>,
     private readonly toggleCase: (rawChar: string) => string
@@ -155,32 +141,4 @@ class MixedCaseArbitrary extends NextArbitrary<string> {
         })
       );
   }
-}
-
-/** @internal */
-function defaultToggleCase(rawChar: string) {
-  const upper = rawChar.toUpperCase();
-  if (upper !== rawChar) return upper;
-  return rawChar.toLowerCase();
-}
-
-/**
- * Randomly switch the case of characters generated by `stringArb` (upper/lower)
- *
- * WARNING:
- * Require bigint support.
- * Under-the-hood the arbitrary relies on bigint to compute the flags that should be toggled or not.
- *
- * @param stringArb - Arbitrary able to build string values
- * @param constraints - Constraints to be applied when computing upper/lower case version
- *
- * @remarks Since 1.17.0
- * @public
- */
-export function mixedCase(stringArb: Arbitrary<string>, constraints?: MixedCaseConstraints): Arbitrary<string> {
-  if (typeof BigInt === 'undefined') {
-    throw new Error(`mixedCase requires BigInt support`);
-  }
-  const toggleCase = (constraints && constraints.toggleCase) || defaultToggleCase;
-  return convertFromNext(new MixedCaseArbitrary(convertToNext(stringArb), toggleCase));
 }
